test(annotation-queues): fold hobby-limit coverage into existing tRPC suite
Move the create-limit cases into annotation-queue-items-trpc.servertest.ts
so they share the existing caller helper instead of a new test file.
